sop2019/requirements.txt
django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
py_eureka_client>=0.7.2,<0.8.0